102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,9 +10,9 @@
 int main(void)
 {
 	int i = 3;
-	int first = 1;
-	int second = 2;
-	int next = first + second;
+	long int first = 1;
+	long int second = 2;
+	long int next = first + second;
 
 	printf("%d, ", first);
 	printf("%d, ", second);
